refactor: move CLI tests to its package folder
