+++ packages/hygieia-cli/__tests__/main.spec.ts
@@ -5,7 +5,7 @@ import {execaCommandSync} from 'execa';
 import fs from 'fs-extra';
 import {afterEach, beforeEach, describe, expect, test} from 'vitest';
 
-import Messages from '../src/messages';
+import Messages from '@/messages';
 
 const CLI_RUNNER = import.meta.env.VITE_TEST_BUILD === '1' ? 'node' : 'tsx';
